requirements.txt
torch>=2.4.0
torchvision>=0.19.0
lightning>=2.6,<2.7
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
requests>=2.31.0
pyyaml>=6.0.0
pytest>=8.0.0
jupyter>=1.0.0
ipywidgets>=8.0.0
gdown>=5.1.0
kaggle>=1.6.0
